+++ lib/application/common/url_launcher_service.dart
@@ -2,7 +2,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 import '../../domain/app_exception.dart';
-import 'state/overlay_loading_provider.dart';
+import 'states/overlay_loading_provider.dart';
 
 /// URL起動サービスプロバイダー
 final urlLauncherServiceProvider = Provider(
